added functionality to close modal when delete event

+++ src/app/_services/events/events.service.ts
@@ -83,7 +83,11 @@ export class EventsService {
    * @description
    * @param id string
    */
-  deleteEvent(id:string) {
+  deleteEvent(id:string): Observable<any> {
+
+    if (environment.mock) {
+      return of<boolean>(true);
+    }
       return this.http.delete<Event>(`${environment.api.salus}/event-tasks/${id}`, httpOptions)
       .pipe(
         tap(data => {
